feat: Change Salary Status to Enum and Update Related Entities
- Added migration to change SalaryProcess.Status from string to int (enum).
- Introduced SalaryStatus enum with values Unpaid and Completed.
- Updated SalaryProcess entity to include PricePerKg and change Status to SalaryStatus.
- Modified SalaryProcess mapping extensions to handle new properties.
- Removed Receipt entity as it was deemed unnecessary.
- Updated repository and service interfaces to reflect changes in method signatures.
- Created FishTypes management pages with CRUD operations.
- Updated sidebar navigation to include FishTypes.
- Enhanced error handling in SalaryProcessService and SalaryProcessingRepository.

+++ FishCalc.Web/Repositories/FishTypeRepository.cs
@@ -16,14 +16,19 @@ public class FishTypeRepository(AppDbContext context) : IFishTypeRepository
 
     }
 
-    public async Task  Delete(FishType fish)
+    public async Task  Delete(int id)
     {
+        var fish = await context.FishTypes.FindAsync(id);
+        if (fish == null)
+        {
+            return;
+        }
        context.FishTypes.Remove(fish);
         await context.SaveChangesAsync();
 
     }
 
-    public async Task<FishType?> GetFishTypeByIdAsync(int id)
+    public async Task<FishType?> GetFishTypeByIdAsync(int? id)
     {
        return await context.FishTypes.Include(x => x.FishPrice)
             .SingleOrDefaultAsync(x => x.Id == id);
